api/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-cloud-storage
numpy
shapely
geopandas
pyproj